requirements.txt
Flask==3.0.3
flask-cors==4.0.1
mysql-connector-python==8.3.0
Werkzeug==3.0.4
resend==0.7.0